requirements.txt
asgiref==3.5.2
Babel==2.10.3
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Django==4.1
django-allauth==0.51.0
django-countries==7.3.2
django-environ==0.9.0
django-js-asset==2.0.0
django-model-utils==4.2.0
django-money==3.0.0
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-richtextfield==1.6.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
install==1.3.5
oauthlib==3.2.0
phonenumbers==8.12.53
Pillow==9.2.0
py-moneyed==2.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
